h1_light_worker/requirements.txt
nats-py>=2.3.0
python-dotenv>=1.0.0
playwright>=1.40.0
